pyproject.toml
[project]
name = "gemCLI"
version = "0.1.0"
description = "Gemini AI for your terminal"
readme = "README.md"
dependencies = [
    "google-genai",
    "python-dotenv",
]

[project.scripts]
gemini = "gemCLI.gemcli:gemini"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gemCLI"]
